Adhere to Module Style Guide
This commit adds a requirement on puppet-lint to the Rakefile so
one can run 'rake lint' against the entire module.

puppet-lint was then used as a guide in correcting various files in
the module to conform to the Puppet Labs Module Style Guide. A few
spec tests were also modified to adjust to changes made for style.

Funtionality has been left unaffected.